requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
python-dotenv==1.0.1
langchain==0.2.16
gigachat==0.1.35
pydantic==2.9.2

